Update patch set 1
Patch Set 1:

Thanks Lusheng,

This file reflects exactly how the committer group looks like which is not the same as the wiki. 

the reason behind this is becuase initially we sent invitations to the people in the wiki and not everyone accepted. Other times, the wiki gets updated but no one notifies me to make the change in the actual group .. 

Please gather the correct information and once Gildas approves we can make the changes again

Patch-set: 1